feat: no need for float64

+++ training/train_clip.py
@@ -1023,7 +1023,7 @@ def main():
         logits -= jax.lax.stop_gradient(logits_max)
         label_logits = jnp.diag(logits)
         log_normalizers = jnp.log(jnp.sum(jnp.exp(logits), axis=axis))
-        return jnp.mean(log_normalizers - label_logits, dtype=jnp.float64)
+        return jnp.mean(log_normalizers - label_logits)
 
     def clip_loss(similarity):
         loss = (cross_entropy(similarity, axis=0) + cross_entropy(similarity, axis=1)) / 2
